requirements.txt
whitenoise==5.2.0
gunicorn==20.0.4
Django==3.0.7
django-heroku==0.3.1
psycopg2==2.8.5
pytz==2019.3
whitenoise==5.2.0
django-cors-headers==3.4.0
django-heroku==0.3.1
djangorestframework==3.11.0
whitenoise==5.2.0
git+git://github.com/mtlynch/ingredient-phrase-tagger.git
